refactor: consolidate prediction into xgboost
